requirements.txt
discord.py==2.1.0
openai==0.26.2
pydub==0.25.1
python-dotenv==0.21.1
requests==2.28.1
youtube_dl==2021.12.17
